Add SIGINT handler
When Ctrl+C was pressed during operation, the script ended with an
error stacktrace. A handler was added to make this forced exit a
little bit more clean.

The handler is disabled at the DO_BUILD stage to allow users to send
SIGINT to the individual build processes without killing the main
process.