build: target: do not require toolchains to clean
The soft clean command (-C) attempted to download toolchains before
executing. This is generally not needed, because toolchains are not
really used for cleaning - it is mostly just removing things with rm.

Additionally, this commit fixes the issue with soft clean on a freshly
cloned repo. EBE attempted to download the toolchains to a non-existing
./bin/ directory.