utils: change the way of passing commands
Added the shell=True argument, so pipes can be used in the build
commands. The use of this option is generally discouraged
when a possibility of executing some arbitrary code exists.
This is not the case with EBE, as anyone with access
to the ini files can also just modify the python scripts.
EBE is also not run with any extra privileges.