target: fix recognizing defconfig subtargets
The scripts were checking whether 'defconfig' was a part of the
subtarget name. This was because the full name of the subtarget
also included the name of the actual target.
For example: "U-Boot defconfig"

This produced some false-positives, for example for the
"generate_defconfig" subtargets.

Instead of looking for a submatch, just split the target string
to extract the subtarget, and do an actual comparison to see if
that is what we are looking for.